Tests/LibWeb/Text/expected/ResizeObserver/observe.txt

Size changed: 200px x 200px
Size changed: 400px x 400px
